requirements.txt
flask
python-dotenv
pillow
together
requests
